+++ CHANGELOG.md
@@ -98,6 +98,7 @@
     printer profiles
   * [#683](https://github.com/foosel/OctoPrint/issues/683) - Fixed heated bed option not being properly displayed in
     printer profiles
+  * [#714](https://github.com/foosel/OctoPrint/issues/714) - Fixed type validation of printer profiles
 * Various fixes without tickets:
   * GCODE viewer now doesn't stumble over completely extrusionless GCODE files
   * Do not deliver the API key on settings API unless user has admin rights